requirements.txt
psycopg2-binary==2.9.3
Django==4.1
flake8==5.0.4
djangorestframework==3.13.1